==> 04.popgen_stats.Rmd <==
---
title: "Population genetic statistics"
output:
  github_document:
    pandoc_args: --webtex
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
library(tidyverse)
library(ggpubr)
library(cowplot)
```

### Observed and expected heterozygosity
We used plink2 to calculate heterozygosity per sample using the command;

```bash
plink2 --vcf Adigi.v2.filtered.vcf.gz --allow-extra-chr --het
```

As part of this calculation PLINK provides the observed and expected heterozygosity (Ho,He) as well as the inbreeding coefficient F which is calculated as $1-\frac{Ho}{He}$


```{r}
het_data <- read_tsv("data/hpc/popgen/plink2.het") %>% 
  rename(sample=`#IID`) %>% 
  filter(sample!="BR_5_121_S125_L004") %>% 
  mutate(location= case_when(
      grepl("^AI",sample) ~ "IN",
      grepl("^BR",sample) ~ "IN",
      grepl("^AR",sample) ~ "NO",
      grepl("^RS",sample) ~ "SO"
          )) %>% 
  mutate(Ho= (OBS_CT-`O(HOM)`)/OBS_CT )  %>% 
  mutate(He= (OBS_CT-`E(HOM)`)/OBS_CT ) %>% 
  pivot_longer(cols = c(F,Ho,He), names_to = "stat", values_to = "value")
```


```{r, fig.height=3.6, fig.retina=2}
source("scripts/color_scheme.R")

f_plot <- het_data %>% 
  filter(stat=="F") %>% 
  ggplot(aes(x=location,y=value)) + 
  geom_boxplot(aes(color=location),alpha=0.5) + 
  geom_point(aes(x=location,color=location), position = position_jitter(w=0.1,h=0)) + 
  theme_pubr() + 
  theme(legend.position = "bottom",legend.title = element_blank(), text = element_text(size=12) , axis.text.x = element_text(size=12)) +
  scale_color_manual(values = myCol) + 
  xlab("") + ylab("Inbreeding Coeffient")


het_plot <- het_data %>% 
  filter(stat!="F") %>% 
  ggplot(aes(x=location,y=value)) + 
  geom_boxplot(aes(fill=stat,color=stat),alpha=0.5) + 
#  geom_point(aes(x=location,color=stat), position = position_jitter(w=0.1,h=0)) + 
  theme_pubr() + 
  theme(legend.position = "bottom",legend.title = element_blank(), text = element_text(size=12) , axis.text.x = element_text(size=12)) +
#  scale_color_manual(values = myCol) + 
  xlab("") + ylab("Heterozygosity")

plot_grid(f_plot,het_plot,labels=c("A","B"))
```

**Figure 1:** Summary of heterozygosity and inbreeding coefficients in each of the populations.  Note that the Inbreeding coefficient (A) captures deviations from expectation (ie between He and Ho) (B).


```{r}
# Plot for saving only
f_plot_out <- het_data %>% 
  filter(stat=="F") %>% 
  ggplot(aes(x=location,y=value)) + 
  geom_boxplot(aes(color=location),alpha=0.5) + 
  #geom_point(aes(x=location,color=location), position = position_jitter(w=0.1,h=0)) + 
  theme_pubr() + 
  theme(legend.position = "none",text = element_text(size=12) , axis.text.x = element_text(size=12)) +
  scale_color_manual(values = myCol) + 
  xlab("") + ylab("Inbreeding Coeffient")

#ggsave(f_plot_out,filename = "figures/inbreeding.png",width = 3,height = 3)
```


### Pairwise Fst values
Genome-wide pairwise Fst values were calculated between inshore, offshore north, and offshore south using vcftools. Here we calculated in a per SNP manner and get the average value and standard deviation of each pair.
```bash
vcftools --gzvcf Adigi.v2.filtered.vcf.gz \
--weir-fst-pop {pop1} \
--weir-fst-pop {pop2} \
--out {pop1}_{pop2}_per_snp_fst
```
The pairwide average Fst values among three *A.digitifera*. 
```{r fst-table,results='asis'}
fst <- "
|                  | Sample size  | Inshore | North Offshore| South Offshore|
|------------------|:------------:|:-------:|:-------------:|:-------------:|
|**Inshore**       |29            |  -      | 0.055933      |  0.058331     |
|**North Offshore**|20            | 0.022377| -             |  0.035717     |
|**South Offshore**|25            | 0.023734| 0.006928      |   -           |
"
cat(fst)
```

**\*Note:** Mean Fst values are presented in the bottom left and standard deviations are shown in the top right of the matrix.


### Nucleotide diversity and Tajima's D in sliding windows

Next, we computed some standard population genomic statistics in sliding windows (10kb window with 2k sliding window). For the Fst calculation, only bi-allelic variant sites are needed, whereas for nucleotide diversity (pi) and Tajima's D, the monomorphic sites are also used. To get a good estimate of missing genotyped sites in each window, we used `bedtools` to extract sites with called genotypes.

**Calculate the number of effective sites in each window**

```bash
#1. create genome bed file
bioawk -c fastx '{print $name"\t1\t"length($seq)}' \
 reference.fa > reference.bed
 
#2 create windows of 10k size
bedtools makewindows -b reference.bed -w 9999 -s 2000 > windows_10k.bed

#3. calculate the percentage of missing sites in every window
bedtools  genomecov -ibam {sample}.bam -d | \
 awk '{if($3<3)print$0}' | \
 awk '{print $1"\t"$2"\t"$2}' |\
 bedtools merge |\
 bedtools intersect -a windows_10k.bed -b - -wao | \
 cut -f1,2,3,7 |\
 awk '{sum[$1"\t"$2"\t"$3]+=$4}END{for(i in sum)print i"\t"sum[i]/10000}' |\
 sort -k1,1 -k2,2n -k3,3n > {pop}.windows.missing.txt
```

**Nucleotide diversity $\theta_{\pi}$**

vcftools was then used to calculate the nucleotide diversity in each population. Next, windows containing greater than 30% uncalled sites were discarded.

```bash
vcftools --gzvcf Adigi.v2.filtered.vcf.gz --keep {pop} \
 --window-pi 10000 --window-pi-step 2000 --out {pop}
 
awk '{if ($4<0.3) print $0}' {pop}.windows.missing.txt |\
 cut -f1,2,3 | grep -Fwf - {pop}.windowed.pi > {pop}.filtered_windowed.pi
```

**Tajima's D**

Similarly, we used `vk tajima` in [VCF-kit](https://vcf-kit.readthedocs.io/en/latest/) to calculate the genome-wide Tajima's D.

```bash
bcftools view -S {pop} Adigi.v2.filtered.vcf.gz |\
 vk tajima 10000 2000 - |\
 sed '1d' |\
 awk '{print $1"\t"$2+1"\t"$3"\t"$6}' > {pop}_1based.td
 
awk '{if($4<0.3)print $0}' {pop}.window.missing.txt |\
cut -f1,2,3 |\
grep -Fwf - {pop}_1based.td > {pop}.filtered_windowed.td
```

```{r pi, cache=TRUE, fig.height=3.6, fig.retina=2}
read_pi <- function(pop) {
 read_tsv(paste0("data/hpc/popgen/",pop,".filtered_windowed.pi"),col_names = c("chr","start","end","n","pi")) %>% add_column(pop=pop) 
}

pi_plot <- rbind(read_pi("inshore"),read_pi("northoffshore"),read_pi("southoffshore")) %>% 
ggplot() + 
  geom_boxplot(aes(x=pop,y=pi,fill=pop),outlier.size = .5) + 
  ggsci::scale_fill_startrek() + 
  theme_test(base_size = 12) +
  theme(legend.position = "none") + 
  labs(x="",y="Pi") + 
  scale_x_discrete(labels=c("IN","NO","SO"))


read_td <- function(pop) {
  read_tsv(paste0("data/hpc/popgen/",pop,".filtered_windowed.td"), col_names = c("chr","start","end","td")) %>% 
    add_column(pop=pop)
}

td_plot <- rbind(read_td("inshore"),read_td("northoffshore"),read_td("southoffshore")) %>% 
ggplot() +
  geom_boxplot(aes(x=pop,y=td,fill=pop),outlier.size = .5) + 
  ggsci::scale_fill_startrek() + 
  theme_test(base_size = 12) + 
  scale_y_continuous(limits = c(-3,2)) +
  theme(legend.position = "none") + 
  labs(x="",y="Tajima's D") + 
  scale_x_discrete(labels=c("IN","NO","SO"))

plot_grid(pi_plot,td_plot)
#ggsave("figures/fig-s11.jpg",height = 2.6,width = 5.6)
```

**Figure: 2** The left boxplot shows the values of $\theta_{\pi}$ and the right plot displays the Tajima's D values in three populations.

### Linkage disequilibrium decay distribution
We used plink to calculated the $r^2$ of every pairs of SNPs in inshore, offshore north, offshore south. Because of unequal sample sizes in three populations, we randomly sample down to 20 individuals in inshore and offshore south.

```bash
shuf -n 20 {pop}.txt |awk '{print $0"\t"$0}' > {pop}_subset.txt

plink --vcf Adigi.v2.filtered.vcf.gz \
  --allow-no-sex --allow-extra-chr --double-id \
  --ld-window 999999 --ld-window-kb 100 --ld-window-r2 0 \
  --out {pop} --thin 0.01 --r2 \
  --keep {pop}_subset.txt
```

```{r ld-decay, fig.align='center',fig.retina=2,cache=TRUE,fig.width=5.2, fig.height=4.6}
if ( !file.exists("data/r_essentials/ld_summary_data.rds")){
  summarise_ld <- function(filename) {
    df <- read_table2(filename) %>% select(-X8) %>% as_tibble %>% 
      mutate(dist = BP_B-BP_A) %>% arrange(dist) %>%
      mutate(distc = cut(dist,seq(min(dist)-1,max(dist)+1,by=200))) %>% 
      group_by(distc) %>% summarise(mean_dist=mean(dist),mean_r2=mean(R2))
    df
  }

  noffshore_ld <- summarise_ld("data/hpc/popgen/northoffshore.ld") %>% add_column(location="NO")
  soffshore_ld <- summarise_ld("data/hpc/popgen/southoffshore.ld") %>% add_column(location="SO")
  inshore_ld <- summarise_ld("data/hpc/popgen/inshore.ld") %>% add_column(location="IN")

  ld_summary_data <- rbind(inshore_ld,noffshore_ld,soffshore_ld)

  write_rds(ld_summary_data,"data/r_essentials/ld_summary_data.rds")
} else {
  ld_summary_data <- read_rds("data/r_essentials/ld_summary_data.rds")
}

ggplot(ld_summary_data,aes(mean_dist/1000,mean_r2)) + geom_point(size = 1,aes(color=location)) + ggsci::scale_color_startrek(labels = c("Inshore", "Offshore North","Offshore South")) +
    theme_classic() + 
  theme(legend.position = c(0.80,0.85),
        legend.title = element_blank()) + labs(x="Pairwise distance in Kb", y="LD (" ~ r^2 ~ ")")
```

**Figure: 3** The plot depicts the interval distance against the LD $r^2$ value. Red, blue, green dots represent values calculated based on inshore, offshore North, offshore South.



### Number of privately fixed SNPs 
Here our objective was to count SNPs with a very high allele frequency (fixed or nearly fixed) in one population, but absent in both other populations. For this we first used PLINK to create separate allele frequency summaries for each population as follows;

```bash
plink2 --vcf Adigi.v2.filtered.vcf.gz --allow-extra-chr --freq 'cols=chrom,pos,ref,alt,altfreq'  --pheno populations.txt --loop-cats 'site' 
```

where the file `populations.txt` has a format like;

```bash
#IID  site
AI_1_001_S102_L004 IN
AI_1_008_merged IN
AI_1_021_S97_L004 IN
```

This generates an `afreq` file for each population with the allele frequencies at each SNP.  We used `paste` to combine these three files and then an awk script [priv.awk](data/hpc/private_alleles/priv.awk).  

```bash
paste plink2.IN.afreq plink2.NO.afreq plink2.SO.afreq | awk '{print $1,$2,$4,$5,$6,$12,$18}' | awk -v mv=1.0 -f priv.awk > 1.0.priv
paste plink2.IN.afreq plink2.NO.afreq plink2.SO.afreq | awk '{print $1,$2,$4,$5,$6,$12,$18}' | awk -v mv=0.9 -f priv.awk > 0.9.priv
```

Note that the awk script accepts an argument `mv` that determines the threshold allele frequency for "fixation".  The script always requires that the allele be absent (freq=0) in the other populations.

When the threshold is set at complete fixation just a single SNP is identified at `BLFC01000154.1 282919` which is fixed in Inshore and absent elsewhere.  This SNP is surrounded by other SNPs with very high (but not quite fixed) frequencies and is located within a gene `adig_s0150.g24` that has a blast hit to Peroxidasin and is on one of our identified putative selective sweeps. See [17.dating_the_selection](17.dating_the_selection.md) for the estimated age of this sweep.

If the threshold is set lower to 0.9 far more SNPs (95) are identified but all are in the inshore population and all are concentrated in just four loci. Two of these fall within putative sweeps and two fall just outside them within 50kb. A full list of these loci is available as the file [0.9.priv](data/hpc/private_alleles/0.9.priv)

### Number of unique SNPs in populations
We counted the number of SNPs specific to each population. These are variants that only exist in one population and are homozygous in other samples.

```bash
bcftools view -S ^{pop}.txt Adigi.v2.filtered.vcf.gz | bcftools filter -i 'AC=0 || AC=AN'
```

|Population    | Sample size | SNP count    |
|--------------|-------------|--------------|
|Inshore       |     29      |  1,235,530   |
|Offshore North|     20      |  1,425,395   |
|Offshore South|     25      |  1,511,531   |
| Total        |     74      |  9,656,554   |
