==> 05.population_structure.Rmd <==
---
title: "Population Structure"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
library(tidyverse)
library(ggsci)
library(stringi)
library(purrr)
library(patchwork)
library(PCAviz)
source(file = "scripts/color_scheme.R")
```

## Data preparation for population structure analysis

In addition to filters already outlined in [02.quality_control](02.quality_control.md), we performed further filtering steps for population structure analysis (PCA, ADMIXTURE) as follows;

1.To avoid potential errors from sequencing, we only included sites with minor allele count greater than or equal to two.

2.Sites in violation of Hardy-Weinberg equilibrium were removed at p-value < 1e-4.

3.To minimize the effect of SNPs in high linkage disequilibrium (LD) region, we filtered out SNPs based on pairwise LD using PLINK v1.9 with the parameter '--indep-pairwise 50 10 0.1'.

### LD pruning

Because plink will only take common chromosome ids like in human, we first created a map file for *Acropora digitifera*. Note that the common workaround of using the flag `--allow-extra-chr` does not work in this instance because downstream analyses (ADMIXTURE) require numeric scaffold ids.

```bash
bcftools view -H Adigi.DPg90gdp3gq30.vcf.gz | \
  cut -f 1| uniq | \
  awk '{print $0"\t"$0 }' > Adigi.DPg90gdp3gq30.chrom-map.txt
```

Next, we converted our vcf file to plink format which included sites with a minor allele count greater than or equal to two and excluded sites with a p-value for exact test for Hardy-Weinberg Equilibrium below 1e-4. Additionally, we pruned SNPs in highly linked regions.

```bash
plink --vcf Adigi.v2.filtered.vcf.gz --out Adigitifera \
  --allow-extra-chr --make-bed --set-missing-var-ids @:# \
  --double-id --geno 0.1 --mac 2 --hwe 0.0001

plink --bfile Adigitifera --indep-pairwise 50 10 0.1 --out Adigitifera \
    --allow-extra-chr --set-missing-var-ids @:#
    
plink --bfile Adigitifera --extract Adigitifera.prune.in \
    --recode12 --allow-extra-chr --out Adigitifera_ldpruned
```
This left a total of **919,129** SNPs for PCA and Admixture analysis

### Principal Component Analysis

We used `smartpca` from the Eigensoft suite for this task. 

`smartpca` has a design flaw that it won't work with chromosome ids other than small numbers (1-22,X,Y). So we used an [in-house script](scripts/fix_mapfile_for_smartpca.py) to fix the map file and run smartpca with default parameters.

```{r smartpca}
pca <- read.table("data/hpc/pca/Adigitifera_smartpca.evec",
                       comment = "#",sep="",
                       col.names =c("sampleid",paste0("PC",1:20),"location"))

sdev_eval <- sqrt(scan("data/hpc/pca/Adigitifera_smartpca.eval")[1:20])
sum_eval <- sum(scan("data/hpc/pca/Adigitifera_smartpca.eval"))
pve <- sdev_eval/sum_eval*100


p1 <- ggplot(pca %>% filter(sampleid!="BR_5_121:BR_5_121_S125_L004") %>% arrange(desc(location)), aes(PC1,PC2,color=location)) +
  geom_point(size=2,alpha=0.8,shape=16)  +
  scale_color_manual(values = my_color) + 
  labs(x=paste0("PC1(", signif(pve[1],3),"%)"), y=paste0("PC2(", signif(pve[2],3),"%)")) + theme_bw() +
  theme(text=element_text(size=12,  face = "bold"),
        legend.text = element_text(size=8,face = "bold"),
        legend.title = element_blank(),
        legend.position = c(0.4,0.85),
        legend.box = "horizontal",
        legend.direction = "horizontal",
        panel.grid = element_blank()) 

p2 <- ggplot(pca %>% filter(sampleid!="BR_5_121:BR_5_121_S125_L004")%>% arrange(desc(location)), aes(PC2,PC3,color=location)) +
  geom_point(size=2,alpha=0.8,shape=16)  +
  scale_color_manual(values = my_color) + 
  labs(x=paste0("PC2(", signif(pve[2],3),"%)"), y=paste0("PC3(", signif(pve[3],3),"%)")) + theme_bw() +
  theme(text=element_text(size=12,  face = "bold"),
        legend.position = "none",
        panel.grid = element_blank()) 

p1 + p2
```

**Figure 1:** Plots depict the first vs second principal components (PC1 vs PC2) and PC2 vs PC3, stratified by sample site location (red-like dots: Inshore, green-like dots from south offshore, blue dots: north offshore).  


Corals from different regions were well separated as inshore, north offshore, and south offshore, except one sample from Beagle reef was clustered with south offshore. In the right plot, inshore samples were tightly congregated compared to two offshore groups. Therefore, our samples from six locations form three clusters which correspond to three geographic regions we called inshore,north offshore, and south offshore.

This analysis revealed that one sample from Beagle Reef (inshore), `BR_5_121` had offshore ancestry despite being collected from inshore.  This sample is the sample that we later identified as being mislabelled (see [here](18.radseq_check.md)). We excluded this sample from subsequent analyses of demography and selection.

### PC loadings
To check whether population structure evident in the PCA was driven by a small number of genomic regions (eg such as inversions) we inspected the PC loading for the first two principle components.

```{r pc_loading,eval=FALSE}
# Very large file. Skip and simply show the plot in RMarkdown rendering
#
snpsigs <- read_table("data/hpc/pca/Adigitifera_smartpca.snpeigs",col_names = c("ID","chr","pos",paste0("PC",(1:20)))) 

hgdp <- pcaviz(dat=pca, sdev = sdev_eval, var = sum_eval, rotation = snpsigs ) 

load_pc1 <- pcaviz_loadingsplot(hgdp,pc.dim = "PC1",min.rank = 0.8,geom.point.params = list(size=0.5, color="grey36")) + labs(x="SNPs",y=paste0("PC",1)) + ggtitle("PC loading plots of SNPs across the genome")

load_pc2 <- pcaviz_loadingsplot(hgdp,pc.dim = "PC2",min.rank = 0.8,geom.point.params = list(size=0.5, color="grey36")) + labs(x="SNPs",y=paste0("PC",2)) 

library(cowplot)
plot_grid(load_pc1,load_pc2)
ggsave("data/r_essentials/pcloadings.png",width = 8,height = 6)
```

```{r}
knitr::include_graphics("data/r_essentials/pcloadings.png")
```


At a high level this seems to suggest that population structure reflects variation across the entire genome rather and is not dominated by a small number of strongly differentiated regions.

### Admixture analysis

Using the unlinked SNPs in plink format, we ran `ADMIXTURE` with default parameters and 10-fold cross-validation. The postulated number of ancestral population K was set from 1 to 6 and results of the cross-validation error were inspected across this range of values of K.

>`ADMIXTURE` will need genotypes coded as 0,1,2 (instead of A,T,G,C).

```bash
for K in {1..6}
do
  admixture --cv=10 Adigitifera_ldpruned.ped $K | tee log.${K}.out
  awk -v K=$K '$1=="CV"{print K, $4}' log.${K}.out >> CV.txt
done
```
The cross-validation error suggests the lowest cross-validation error is with K=1 which is a common situation when population differentiation is subtle. Since our PCA results clearly pointed toward at least K=3 we then checked admixture proportions for K=2 and 3.  

```{r admixture, fig.width=5.2, fig.height=6,fig.retina=2,out.extra='style="float:left"'}
samples <- scan("data/hpc/admixture/sample_order.txt", character())
read_Q <- function(filename) {
  tb <- read_table(filename,col_names = FALSE) %>% 
  add_column(samples=samples) %>% 
  mutate(pop=str_split(samples,"_") %>% map_chr(first)) %>% 
  gather(cluster,proportion,-samples,-pop)
  
  tb$samples <- factor(tb$samples, levels = tb %>% 
                         arrange(pop) %>% pull(samples) %>% unique())
  tb$pop <- factor(tb$pop,levels = c("AI","BR","AR","RS1","RS2","RS3"))
  
  tb$cluster <- factor(tb$cluster, levels=rev(unique(tb$cluster)))
  
  tb
}
pop.labs <- c("AI","BR","AR","RS1","RS2","RS3")
names(pop.labs) <- c("AI","BR","AR","RS1","RS2","RS3")
p1 <- ggplot(read_Q("data/hpc/admixture/Adigitifera_ldpruned.2.Q")%>% filter(samples!="BR_5_121_S125_L004"),aes(x=samples)) + geom_col(aes(group=samples,y=proportion,fill=cluster),show.legend = FALSE) + 
  theme_minimal() + 
  facet_grid(~pop,switch = "x",scales = "free",space = "free") +
  scale_color_brewer(palette = "Set2") +
  theme(axis.text.x = element_blank(),panel.grid = element_blank(),strip.text.x = element_blank())+
  labs(y="Ancestry (K=2)", x="") 

p2<-ggplot(read_Q("data/hpc/admixture/Adigitifera_ldpruned.3.Q")%>% filter(samples!="BR_5_121_S125_L004"),aes(x=samples)) + geom_col(aes(group=samples,y=proportion,fill=cluster),show.legend = FALSE) + theme_minimal() +
  facet_grid(~pop,switch = "x",scales = "free",space = "free",
             labeller = labeller(pop=pop.labs)) +
  scale_fill_brewer(palette = "Set2",labels= c("X1"="South Offshore","X2"="North Offshore", "X3" = "Inshore")) +
  theme(axis.text.x = element_blank(),panel.grid = element_blank())+
  labs(y="Ancestry (K=3)", x="") 

p3<-ggplot(read_Q("data/hpc/admixture/Adigitifera_ldpruned.4.Q")%>% filter(samples!="BR_5_121_S125_L004"),aes(x=samples)) + geom_col(aes(group=samples,y=proportion,fill=cluster),show.legend = FALSE) + theme_minimal() +
  facet_grid(~pop,switch = "x",scales = "free",space = "free",
             labeller = labeller(pop=pop.labs)) +
  scale_fill_brewer(palette = "Set2") +
  theme(axis.text.x = element_blank(),panel.grid = element_blank())+
  labs(y="Ancestry (K=4)", x="") 

p4<-ggplot(read_Q("data/hpc/admixture/Adigitifera_ldpruned.5.Q")%>% filter(samples!="BR_5_121_S125_L004"),aes(x=samples)) + geom_col(aes(group=samples,y=proportion,fill=cluster),show.legend = FALSE) + theme_minimal() +
  facet_grid(~pop,switch = "x",scales = "free",space = "free",
             labeller = labeller(pop=pop.labs)) +
  scale_fill_brewer(palette = "Set2") +
  theme(axis.text.x = element_blank(),panel.grid = element_blank())+
  labs(y="Ancestry (K=5)", x="") 

p1/p2/p3/p4
```


```{r}
# For inclusion in Figure 1
pop_order <- c("AR"=1,"AI"=2,"BR"=3,"RS1"=4,"RS2"=5,"RS3"=6)
cluster_cols <- myCol
names(myCol) <- c("X1","X2","X3")

admix_vp <- read_Q("data/hpc/admixture/Adigitifera_ldpruned.3.Q") %>% 
  filter(!grepl("BR_5_121",samples)) %>% 
  mutate(order = pop_order[pop]) %>% 
    ggplot(aes(x=samples)) + 
    geom_col(aes(group=samples,y=proportion,fill=cluster),show.legend = FALSE) + theme_minimal() +
    coord_flip() +
    facet_wrap(~factor(pop,levels = names(pop_order)),
               scales = "free_y",
              labeller = labeller(pop=pop.labs),
              ncol = 1,
              strip.position = "left") +
  scale_fill_manual(values = cluster_cols) +
  theme(axis.text.y = element_blank(),panel.grid = element_blank())+
  labs(y="Ancestry (K=3)", x="") 


#ggsave(admix_vp,filename = "figures/admixture_vertical.png",width = 2,height = 6)
```



```{r cv, fig.retina=2, fig.width=3.4, fig.height=3.8,out.extra='style="float:left"'}
cv <- read_delim("data/hpc/admixture/CV.txt", col_names = c("CV","value"),delim = " ")
ggplot(cv,aes(x=CV,y=value)) + geom_line() + geom_point(shape=21,fill="white",alpha=1) +
  theme_bw() +
  theme(panel.grid = element_blank()) +
  labs(x="K",y="Cross-validation error")
```
