==> 20.inshore_candidate_genes.Rmd <==
---
title: "Genes under selection in inshore reefs"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
library(tidyverse)
library(ggrepel)
```

We extract top 1% outlier in iHS, XP-EHH, and XP-nSL and made a statistic of all haplotype-based selection scan using `multiIntersectBed`. In the interected bed file, it displays the all regions that were identified as significantly under selection in as least one method and the occurrences of them.

```bash
multiIntersectBed -i {pop}*.percentile1.bed -header > {pop}.multiIntersect.bed
```
For these windows, we checked several signals such as XP-EHH, XP-nSL, iHS and delta-iHH, delta-DAF, and Fst.


```{r uniprot-table}
bl6_cols <- c("qaccver","saccver","pident","length","mismatch","gapopen","qstart","qend","sstart","send","evalue","bitscore")
blastp <- read_tsv("data/hpc/annotation/Adigitifera_uniprot_blastp.outfmt6", col_names = bl6_cols) %>% 
  select(adi_id=qaccver, uniprot_id=saccver,evalue) %>% add_column(method="blastp") %>% mutate(geneid=str_remove(adi_id,'\\.t\\d'))

blastx <- read_tsv("data/hpc/annotation/Adigitifera_uniprot_blastx.outfmt6", col_names = bl6_cols) %>% 
  select(adi_id=qaccver, uniprot_id=saccver,evalue) %>% add_column(method="blastx") %>% mutate(geneid=str_remove(adi_id,'\\.t\\d'))

uniprotkb_tab <- read_tsv("data/hpc/annotation/uniprot-yourlist.tab") %>% select(-Status)

uniprot_gene_annot <- rbind(blastp,blastx) %>% group_by(geneid) %>% dplyr::top_n(1,dplyr::desc(evalue)) %>% group_by(geneid,evalue) %>% filter(row_number()==1) %>% 
  left_join(uniprotkb_tab,by=c("uniprot_id"="Entry")) %>%
  ungroup %>% 
  select(geneid,uniprot_id,entryname="Entry name",
         genename="Gene names",go="Gene ontology IDs",kegg="Cross-reference (KEGG)") %>% mutate(genename=ifelse(is.na(genename),entryname,genename))

adi_id_genename <- uniprot_gene_annot %>% select(geneid,entryname)
```


### Gene candidates in Inshore
Combine all the significant regions in iHS, XP-EHH, and XP-nSL scans. There are 9 windows (including 33 genes) identified as under selection in all 5 comparisons in inshore reefs (iHS, XP-EHH and XP-nSL against north offshore and south offshore reefs).
```{r}

genes_in_windows <- read_tsv("data/hpc/selection/results/inshore.multiIntersect.genes.bed",
                        col_names = c("chrom","start","end","adi_id")) %>% 
                    left_join(adi_id_genename,by=c("adi_id"="geneid")) %>% mutate(entryname=case_when(!is.na(entryname) ~ entryname, is.na(entryname) ~ adi_id)) %>% 
                    group_by(chrom,start,end) %>% summarise(genes=paste(entryname %>% unique,collapse = "; ")) 

windows_to_genes_tbl<- read_tsv("data/hpc/selection/results/inshore.multiIntersect.bed") %>% mutate(chrom=str_remove(chrom,"\\.1")) %>% select(chrom,start,end,num) %>% left_join(genes_in_windows,by=c("chrom","start","end")) %>% 
  mutate(genes=ifelse(is.na(genes),"no gene in this window",genes)) %>% 
  arrange(desc(num)) %>% 
  rename(Scaffold_ID=chrom, Start=start, End=end, "Significant in #"=num, "Genes in this window"=genes)

knitr::kable(windows_to_genes_tbl)

```
We then extracted the norm values 
```{r plot-windows,fig.height=9,fig.width=15}
all_genes <- read_tsv("data/hpc/selection/windows/intersect_genes.bed",col_names = c("scaff","start","end","scaff1","gene_start","gene_end","geneid")) %>% select(-scaff1) %>%  left_join(adi_id_genename) %>% mutate(genename=ifelse(is.na(genename),geneid,genename)) %>% 
  mutate(location=paste(scaff,paste(start,end,sep="-"),sep = ":")) %>% 
  select(location,genename,gene_start,gene_end)

bed <- read_tsv("data/hpc/selection/windows/intersect.bed",col_names = c("scaff","start","end"))

ihs <- read_tsv("data/hpc/selection/windows/intersect_windows.ihs.norm",col_names = F) %>% select(X1,X2,X7) %>% 
  rename("scaff"=X1, "pos"=X2,"value"=X7) %>% add_column(type="ihs")

xpehh_no <- read_tsv("data/hpc/selection/windows/intersect_windows.xpehh_no.norm",col_names = F) %>% select(X1,X2,X9) %>% 
  rename("scaff"=X1, "pos"=X2,"value"=X9) %>% add_column(type="xpehh_no")

xpehh_so <- read_tsv("data/hpc/selection/windows/intersect_windows.xpehh_so.norm",col_names = F) %>% select(X1,X2,X9) %>% 
  rename("scaff"=X1, "pos"=X2,"value"=X9) %>% add_column(type="xpehh_so")

xpnsl_no <- read_tsv("data/hpc/selection/windows/intersect_windows.xpnsl_no.norm",col_names = F) %>% select(X1,X2,X9) %>% 
  rename("scaff"=X1, "pos"=X2,"value"=X9) %>% add_column(type="xpnsl_no")

xpnsl_so <- read_tsv("data/hpc/selection/windows/intersect_windows.xpnsl_so.norm",col_names = F) %>% select(X1,X2,X9) %>% 
  rename("scaff"=X1, "pos"=X2,"value"=X9) %>% add_column(type="xpnsl_so")

rbind(ihs,xpehh_no,xpehh_so,xpnsl_no,xpnsl_so) %>% left_join(bed) %>% filter(pos>=start & pos<=end) %>% mutate(location=paste(scaff,paste(start,end,sep="-"),sep = ":")) %>% select(location, pos, value, type) %>% 
  ggplot(aes(x=pos/1000,y=value)) + geom_point(size=1) +  
  geom_segment(data=all_genes %>% add_column(type="xpnsl_so"),aes(x=gene_start/1000,xend=gene_end/1000,y=-2.5,yend=-2.5))+ 
  geom_text_repel(data=all_genes %>% add_column(type="xpnsl_so"),aes(x=(gene_start+gene_end)/2000,y=-2,label=genename),size=2.5)+
  facet_grid(type~location, scales = "free")  + theme_classic()
  

```

This plot is very difficult to see. We will plot the combined signals for each window:
```{r}
fst <- read_tsv("data/hpc/selection/windows/intersect_windows.fst_inshore_vs_offshore.weir.fst") %>% rename("scaff"=CHROM, "pos"=POS,"value"=WEIR_AND_COCKERHAM_FST) %>% add_column(type="fst") %>% mutate(scaff=str_remove(scaff,"\\.1"))

all_selscan <- rbind(ihs,xpehh_no,xpehh_so,xpnsl_no,xpnsl_so,fst)  %>% left_join(bed) %>% filter(pos>=start & pos<=end) %>% mutate(location=paste(scaff,paste(start,end,sep="-"),sep = ":")) %>% select(location, pos, value, type) 

window_plot <- function(region) {
  all_selscan %>% filter(location==region) %>% 
  ggplot(aes(x=pos/1000,y=value)) + geom_point(size=.5) +  
    #geom_segment(data=all_genes %>% add_column(type="xpnsl_so"),aes(x=gene_start/1000,xend=gene_end/1000,y=-2.5,yend=-2.5))+ 
    #geom_text_repel(data=all_genes %>% add_column(type="xpnsl_so"),aes(x=(gene_start+gene_end)/2000,y=-2,label=genename),size=2.5)+
    facet_grid(type~.,scales = "free")  + theme_classic()
}
window_plot("BLFC01000074:100001-150001")
window_plot("BLFC01000326:1550001-1600001")
window_plot("BLFC01000326:1750001-1800001")
window_plot("BLFC01000653:2050001-2100001")

```
