==> 24.nobottle_sim.Rmd <==
---
title: "Simulated data under a model without a bottleneck"
bibliography: bibliography.bib
output:
  github_document:
    pandoc_args: --webtex
---

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = FALSE,message=FALSE,warning = FALSE)
library(tidyverse)
library(ggpubr)
library(cowplot)
library(ggpubr)
library(ggsci)
```

Analyses with SMC++ and fastsimcoal both indicated the presence of a strong bottleneck (most likely founder effect) in all three populations.  These bottleneck effects would have contributed to drift, which, combined with low migration rates would have resulted in the clear divergence observed in population structure analyses.  Here we used fastsimcoal analyses to disentangle these effects.  To do this we constructed a model (called `no_bott`) that was based on the best fitting model, but, where all changes in Ne were removed, effectively keeping the population size the same as the ancestral population with subdivision on splits.  This model is detailed in [3.out.no_bott.est](data/hpc/fastsimcoal/no_bott/3.out.no_bott.est) and [3.out.no_bott.tpl](data/hpc/fastsimcoal/no_bott/3.out.no_bott.tpl)

To work out input parameters for this model we took average Ne values from the 95% confidence interval ranges obtained from the best fitting IMc model.  Then, assuming that the overall population maintains its ancestral size we apportioned Ne into IN, NO and SO populations so that their final values matched the relative sizes in the most recent epoch in IMc.  

Simulations with this model were generated using

```bash
fsc2705 -t 3.out.no_bott.tpl -n 1 -e 3.out.no_bott.est  -E 10 -G x -c 0 -k 10000000
```

After converting to `vcf` they were used to calculate genome-wide Fst values between all pairs with plink2. This was done separately for each independent simulation.

```bash
plink2 --vcf ${f} --fst site --pheno phenotypes.txt --allow-extra-chr --out fst/${s2}_no_bott.fst
```

The boxplot below shows that these Fst values are much smaller than those for the full model (`IMc`) that includes population size changes. Thus indicating a strong contribution from the bottleneck to drift and population divergence. 

```{r}
read_fst <- function(path){
  read_tsv(path,show_col_types=FALSE) %>% 
    unite("Pair",`#POP1`,POP2)
}

if (!file.exists("data/r_essentials/24_fst.rds")){

  fst_sum_nb <- list.files("data/hpc/fastsimcoal/no_bott/3.out.no_bott/fst/","*.summary", full.names = TRUE) %>% 
    map_dfr(read_fst) %>% 
    add_column(model="no_bott")

  fst_sum_sc <- list.files("data/hpc/fastsimcoal/growth_rate_SC/3.out.growth_rate_SC/fst/","*.summary", full.names = TRUE) %>% 
    map_dfr(read_fst) %>% 
    add_column(model="IMc")

  fst_sum <- rbind(fst_sum_nb,fst_sum_sc) 
  write_rds(fst_sum,"data/r_essentials/24_fst.rds")
} else {
  fst_sum <- read_rds("data/r_essentials/24_fst.rds")  
}
  
p_fst <- fst_sum %>% 
  ggplot(aes(x=Pair)) + 
  geom_boxplot(aes(y=HUDSON_FST,fill=model)) + 
  scale_fill_discrete(labels = c("IMc","No Bottleneck")) +
  theme_pubclean() +
  theme(legend.title = element_blank()) +
  xlab("Population Pair") + ylab("Fst")

nobott_pca <- read_tsv("data/hpc/fastsimcoal/no_bott/3.out.no_bott/pca/3.out.no_bott_1_1.eigenvec") %>% 
  tidyr::extract(`#IID`,"Location",regex = "([^_]+)",remove=FALSE)

p_pca <- nobott_pca %>% 
  ggplot(aes(x=PC1,y=PC2)) + 
  geom_point(aes(color=Location)) +
  theme_pubclean() +
  scale_color_startrek()

plot_grid(p_fst,p_pca,nrow=1,labels = c("A","B"))
ggsave("figures/fig-s17.png",width = 6.5,height = 4)
```
